Extract no-op action
